// src/app/demo/model/Marcador.ts
export interface Marcador{
    CodigoMarcador: string;   // Código único para el marcador
    DescripcionMarcador: string; // Descripción del marcador
    IP: string;                // Dirección IP del marcador
    CodigoEquipo: string;      // Código del equipo al que pertenece
    NombreEquipo: string;      // Nombre del equipo al que pertenece
    id:string;
}